Updated to MAME 0.104u3 core

+++ mess/tools/imgtool/stream.h
@@ -21,7 +21,7 @@ void stream_close(imgtool_stream *stream);
 size_t stream_read(imgtool_stream *stream, void *buf, size_t sz);
 size_t stream_write(imgtool_stream *stream, const void *buf, size_t sz);
 UINT64 stream_size(imgtool_stream *stream);
-int stream_seek(imgtool_stream *stream, size_t pos, int where);
+int stream_seek(imgtool_stream *stream, INT64 pos, int where);
 size_t stream_tell(imgtool_stream *stream);
 void stream_clear(imgtool_stream *stream);
 void *stream_getptr(imgtool_stream *stream);
